precision(coordinates): the path is f64 from the input file to the flush
A coordinate is read as f64, travels the wire and the write-ahead log as f64, and
reaches the quantiser at the width the caller sent it. Nothing narrows on the way.

This closes a defect that exists today rather than preparing for one that might:
`tessera_spatial::shape` is f64 throughout while the point path was f32, so a shape
resolves 2^32 steps per axis where a point resolved 2^24. On the Overture corpus three
of 17,551 divisions hold a place that is inside at its source coordinates and outside at
its stored position. The same asymmetry decides *cells* rather than residuals at any
frame past roughly zoom offset 8, where there is less than one f32 step per cell — and
nothing in any report can see that it did, the quantiser having been handed a value the
file did not hold.

It costs nothing. The Overture and GBIF inputs already store `double`; the build holds
coordinates only in a 65,536-row decode batch; and the stored form does not change at
all — a position is still a 32-bit fixed-point pair against the frame, and no float
reaches the bundle.

**Both float widths are accepted on input and the narrower is widened**, on the read side
and on the wire alike. A whole-world frame is served perfectly well by f32 coordinates and
a corpus emitting them should not have to double the size of its two largest columns to be
read; a sub-square frame needs f64 and the caller supplies it. Precision is a property of
the corpus rather than of the release. The other direction does not happen: an attribute's
width is *declared*, so narrowing is what the declaration asked for, whereas a coordinate's
width is nobody's request.

`WAL_VERSION` moves to 16 and a log at 15 is refused rather than migrated. A field's type
changed, not a variant or a field — so a version-15 record has the right length and the
right CRC, and four of the eight bytes a coordinate now takes come out of whatever field
follows it. The row would land somewhere on the grid instead of failing. No deployment
holds a log (decision 0048), so the version is the whole guard.

`tessera_corpus` widens the type of its generated coordinates without changing the
arithmetic — the draw still rounds through f32 and is stored wider — because a wider draw
would be a different corpus rather than a more precise one, and would move every fixture
position and churn every golden digest.

Deliberately untouched: the TypeScript client's Float32Array vertex buffers, which are GPU
data downstream of quantisation and correctly f32; and the F32 attribute-column variants,
which are a declared scalar width rather than a coordinate.

`contracts.md` r47 states the ingest schema's two accepted widths. No projection is
involved: this is a width, not a transform.

Tested, each case a pair of positions rather than a type signature:

- the same rows reached by a build and by an ingest place every point in the same cell,
  at a frame 1/4096 of the coordinate range where one f32 step spans sixteen cells;
- a float32 points file reads to the positions it always did;
- two positions inside one f32 step land in different cells;
- a point inside a polygon 8e-4 units across is inside it at its stored position, where
  an f32 position misses;
- an ingest body carries either width, and a float64 coordinate outside the extent is
  refused rather than rounded onto its inclusive maximum;
- the log round-trips a coordinate no f32 holds, and a log at version 15 is refused —
  against a *complete* header, the six-byte fixture beside it having been refused on its
  length before the version was ever compared.

+++ crates/tessera-bench/src/arms/ingest.rs
@@ -368,8 +368,8 @@ pub(crate) fn synth_rows(
                 external_id: Some(format!("bench-{n}").into_bytes()),
                 view: "s0".to_string(),
                 descriptors: descriptors.to_vec(),
-                x: ((n * 37) % 65536) as f32,
-                y: ((n * 53) % 65536) as f32,
+                x: ((n * 37) % 65536) as f64,
+                y: ((n * 53) % 65536) as f64,
                 scalars: Vec::new(),
                 terms: terms.to_vec(),
             }
@@ -1081,8 +1081,8 @@ fn rate_rows(
                 external_id: Some(format!("rate-{n}").into_bytes()),
                 view: "s0".to_string(),
                 descriptors: picks.iter().map(|&k| pool_descriptors[k].clone()).collect(),
-                x: ((n * 37) % 65536) as f32,
-                y: ((n * 53) % 65536) as f32,
+                x: ((n * 37) % 65536) as f64,
+                y: ((n * 53) % 65536) as f64,
                 scalars: Vec::new(),
                 terms: picks.iter().map(|&k| pool_terms[k]).collect(),
             }
