+++ crates/tessera-engine/tests/dict_generation.rs
@@ -215,8 +215,8 @@ fn a_credential_re_presented_after_a_promoting_flush_sees_the_promoted_descripto
             external_id: Some(external.as_bytes().to_vec()),
             view: "s0".to_string(),
             descriptors: vec![b"novel".to_vec()],
-            x: 10.0 + i as f32,
-            y: 10.0 + i as f32,
+            x: 10.0 + i as f64,
+            y: 10.0 + i as f64,
             scalars: Vec::new(),
             terms: engine.resolve_terms(&[b"novel".to_vec()]),
         };
@@ -328,8 +328,8 @@ fn a_background_refresh_does_not_poison_a_later_authorise_of_the_same_credential
             external_id: Some(external.as_bytes().to_vec()),
             view: "s0".to_string(),
             descriptors: vec![b"novel".to_vec()],
-            x: 10.0 + i as f32,
-            y: 10.0 + i as f32,
+            x: 10.0 + i as f64,
+            y: 10.0 + i as f64,
             scalars: Vec::new(),
             terms: engine.resolve_terms(&[b"novel".to_vec()]),
         };
